requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy[asyncio]
psycopg[binary,pool]
aiosqlite
tenacity
PyYAML
jsonschema
openai>=1.40.0
python-dotenv
passlib[bcrypt]
